+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org)
 
-## 3.0.10-dev
+## 3.1.0 - 2022-10-17
+
+- Added new cache-hit/miss ratio in the cache commands
+- Fix deprecation notice in dependency on php 8.1
 
 ## 3.0.9 - 2022-10-06